+++ Project/frontend/src/account/Profile.jsx
@@ -1,93 +1,85 @@
-import { useState, useRef } from 'react';
-import { useSelector } from 'react-redux'
+import React, { useState, useEffect } from 'react';
 import axios from "axios";
-import {Cookies} from 'react-cookie'
-import { useEffect } from 'react';
-export { ProfileCreate, ProfileDetail }
+import { useSelector, useDispatch } from 'react-redux';
+import { user } from '../redux/user';
+import { useParams } from 'react-router-dom';
+export { Profile };
 
-function ProfileCreate() {
+function Profile() {
     const [info, setInfo] = useState({
-        nickname: '',
-        profileImg: "https://cdn.pixabay.com/photo/2015/10/05/22/37/blank-profile-picture-973460_1280.png",
-        backImg: '',
-        self_intro: ''
+        id: '',
+        nickname: '', 
+        background_image: '',
+        profile_image: '',
+        self_intro: '',
+        user: '',
+        slug: ''
     });
-    const fileInput = useRef(null)
-    const {nickname, profileImg, backImg, self_intro} = info; // 비구조화할당
 
-    const onChange = e => {
-        const {value,name} = e.target;
-        setInfo({
-            ...info,
-            [name]:value
-        });
-    };
-
-    return (
-        <div className="ProfileCreate" style={{textAlign:"center"}}>
-            프로필생성창
-            <div className="ProfileImgBox">
-                <input 
-                    type="file" 
-                    accept="image/jpg, image/jpeg, image/png" 
-                    ref={fileInput}
-                    // style={{display: 'none'}}
-                />
-            </div>
-            <div className="BackImgBox">
-                <input 
-                    type="file" 
-                    accept="image/jpg, image/jpeg, image/png" 
-                    ref={fileInput}
-                    // style={{display: 'none'}}
-                />
-            </div>
-            <input type="text" name="nickname" value={nickname} onChange={onChange} placeholder="nickname"/>
-            <input type="text" name="self_intro" value={self_intro} onChange={onChange} placeholder="self_intro"/>
-            <div className='profile-preview'>
-                <img src={profileImg} alt="profile-preview"  width={100} height={100} />
-            </div>
-            <div className='background-preview'>
-                <img src={backImg} alt="background-preview" />
-            </div>
-        </div>
-    );
-}
-
-function ProfileDetail() {
-    const user = useSelector((state) => state.user.value)
-    const cookies = new Cookies()
-    const testtoken = 'eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJpZCI6MSwiZXhwIjoxNjcwMzkzMjQyfQ.4ttvGfLbXhQZadD03geY9BrrRAblOfBn6SkU5q6u9a0'
-    useEffect(() => {
-        const token = cookies.get('jwt');
-        console.log('test:'+token);
-        
-        axios.get('/accounts/profile', {
+    let { profileUsername } = useParams();
+    profileUsername = localStorage.getItem('username');
+    const authentication = useSelector((state) => state.authenticate.value) //authentication.isAuthenticated 사용
+    const userProfile = useSelector((state) => state.user.value)
+    const dispatch = useDispatch()
+    const token = localStorage.getItem('token');
+    
+    useEffect(()=>{
+        axios.get(`/accounts/profile/detail/${profileUsername}`, {
             headers: {
-                Authorization: `token ${testtoken}`
-            }})
-            .then(res => {
-                console.log(res);
-            })
-            .catch(error => console.log(error))
-    })
+                'Authorization': `token ${token}`,
+                'Content-Type':'application/json'
+            },  
+        })
+        .then(function(res){
+            setInfo(res.data)
+        })
+        .catch(error=>console.log(error))
+    }, [profileUsername, token]);
+
+
+    useEffect(()=>{
+        dispatch(user({
+            id: info.id,
+            nickname: info.nickname, 
+            background_image: info.background_image,
+            profile_image: info.profile_image,
+            self_intro: info.self_intro,
+            user: info.user,
+            slug: info.slug
+        }))
+    }, [info, dispatch])
+    const page_hosturl = 'https://'+window.location.hostname
 
     return (
-        <div className='ProfileDetail'>
-            <p>프로필 화면 만들거다</p>
-            <div>
-                <p>아마 유저네임: {user.username}</p>
-                <p>아마 닉네임: {user.nickname}</p>
-                <p>아마 자기소개: {user.selfIntro}</p>
-                <div>
-                    <p>아마사진: </p>
-                    <div className='profile-preview'>
-                        <img src="https://cdn.pixabay.com/photo/2015/10/05/22/37/blank-profile-picture-973460_1280.png" alt="profile-preview"  width={100} height={100} />
+        <div className="Profile">
+            <div className="Profile-container">
+                <div className='background-image'>
+                    <img src={`${page_hosturl}${userProfile.profile_image}`} alt="profile-preview"  width={100} height={100} />
+                </div>
+                <div className='profile-image'>
+                    <img src={`${page_hosturl}${userProfile.profile_image}`} alt="profile-preview"  width={100} height={100} />
+                </div>
+                <button className='profile-edit'>프로필 수정</button>
+                <div className='profile-content'>
+                    <p className='username'>{profileUsername}</p>
+                    <p className='nickname'>{`@${userProfile.nickname}`}</p>
+                    <p className='self-intro'>{userProfile.self_intro}</p>
+                    <div className='profile-nuri-inform'>
+                        <div className='count diary-num'>
+                            <p className='num'>132</p>
+                            <p className='what'>일기</p>
+                        </div>
+                        <div className='count diary-num'>
+                            <p className='num'>46</p>
+                            <p className='what'>팔로잉</p>
+                        </div>
+                        <div className='count diary-num'>
+                            <p className='num'>329</p>
+                            <p className='what'>팔로워</p>
+                        </div>
                     </div>
                 </div>
-                <p>기타 등등</p>
             </div>
-
         </div>
     );
 }
